Fixed problems

+++ Researcher.cs
@@ -43,7 +43,7 @@ namespace A2SDD
 
         public List<Publication> Publications { get; set; }
 
-        public float Performance { get; set; }
+        public double Performance { get; set; }
 
         
         public ObservableCollection<Position> GetPositions()
